+++ tests/testthat/test-class_database.R
@@ -367,3 +367,14 @@ tar_test("database queue", {
   lines <- readLines(db$path)
   expect_equal(lines, c("x", "y"))
 })
+
+tar_test("compare_working_directories()", {
+  on.exit(tar_runtime$unset_working_directory())
+  tar_runtime$set_working_directory(getwd())
+  expect_silent(compare_working_directories())
+  tar_runtime$set_working_directory("..")
+  expect_error(
+    compare_working_directories(),
+    class = "tar_condition_run"
+  )
+})